requirements.txt
absl-py==0.6.0
astor==0.7.1
cycler==0.10.0
Django==2.1.3
djangorestframework==3.9.0
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
imageai==2.0.2
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.0
numpy==1.15.3
opencv-python==3.4.3.18
Pillow==5.3.0
protobuf==3.6.1
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.7
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
Werkzeug==0.14.1
